+++ FoodOrderingApp-api/src/main/java/com/upgrad/FoodOrderingApp/api/controller/AddressController.java
@@ -9,8 +9,8 @@ import com.upgrad.FoodOrderingApp.api.model.SaveAddressRequest;
 import com.upgrad.FoodOrderingApp.api.model.SaveAddressResponse;
 import com.upgrad.FoodOrderingApp.api.model.StatesList;
 import com.upgrad.FoodOrderingApp.api.model.StatesListResponse;
-import com.upgrad.FoodOrderingApp.service.business.AddressService;
-import com.upgrad.FoodOrderingApp.service.business.CustomerService;
+import com.upgrad.FoodOrderingApp.service.businness.AddressService;
+import com.upgrad.FoodOrderingApp.service.businness.CustomerService;
 import com.upgrad.FoodOrderingApp.service.entity.AddressEntity;
 import com.upgrad.FoodOrderingApp.service.entity.CustomerAddressEntity;
 import com.upgrad.FoodOrderingApp.service.entity.CustomerEntity;
@@ -18,6 +18,7 @@ import com.upgrad.FoodOrderingApp.service.entity.StateEntity;
 import com.upgrad.FoodOrderingApp.service.exception.AddressNotFoundException;
 import com.upgrad.FoodOrderingApp.service.exception.AuthorizationFailedException;
 import com.upgrad.FoodOrderingApp.service.exception.SaveAddressException;
+import com.upgrad.FoodOrderingApp.service.util.UtilityProvider;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
@@ -56,7 +57,7 @@ public class AddressController {
     public ResponseEntity<SaveAddressResponse> saveAddress(@RequestHeader("authorization") final String authorization, @RequestBody(required = false)SaveAddressRequest saveAddressRequest)throws AuthorizationFailedException, AddressNotFoundException, SaveAddressException {
 
         //Access the accessToken from the request Header
-      String accessToken = customerService.getBearerAccessToken(authorization);
+      String accessToken = UtilityProvider.decodeBearerToken(authorization);
 
         //Calls customerService getCustomer Method to check the validity of the customer.this methods returns the customerEntity  to be updated  with address.
         CustomerEntity customerEntity = customerService.getCustomer(accessToken);
@@ -95,7 +96,7 @@ public class AddressController {
     public ResponseEntity<AddressListResponse> getAllSavedAddress(@RequestHeader("authorization")final String authorization)throws AuthorizationFailedException{
 
         //Access the accessToken from the request Header
-      String accessToken = customerService.getBearerAccessToken(authorization);
+      String accessToken = UtilityProvider.decodeBearerToken(authorization);
 
         //Calls customerService getCustomer Method to check the validity of the customer.this methods returns the customerEntity  to be get all address.
         CustomerEntity customerEntity = customerService.getCustomer(accessToken);
@@ -137,7 +138,7 @@ public class AddressController {
       throws AuthorizationFailedException, AddressNotFoundException {
 
         //Access the accessToken from the request Header
-    String accessToken = customerService.getBearerAccessToken(authorization);
+    String accessToken = UtilityProvider.decodeBearerToken(authorization);
 
         //Calls customerService getCustomer Method to check the validity of the customer.this methods returns the customerEntity of whose address be deleted.
         CustomerEntity customerEntity = customerService.getCustomer(accessToken);
