+++ FoodOrderingApp-api/src/test/java/com/upgrad/FoodOrderingApp/api/controller/CategoryControllerTest.java
@@ -13,7 +13,7 @@
  import com.fasterxml.jackson.databind.ObjectMapper;
  import com.upgrad.FoodOrderingApp.api.model.CategoriesListResponse;
  import com.upgrad.FoodOrderingApp.api.model.CategoryDetailsResponse;
- import com.upgrad.FoodOrderingApp.service.business.CategoryService;
+ import com.upgrad.FoodOrderingApp.service.businness.CategoryService;
  import com.upgrad.FoodOrderingApp.service.entity.CategoryEntity;
  import com.upgrad.FoodOrderingApp.service.entity.ItemEntity;
  import com.upgrad.FoodOrderingApp.service.exception.CategoryNotFoundException;
