+++ FoodOrderingApp-api/src/test/java/com/upgrad/FoodOrderingApp/api/controller/CustomerControllerTest.java
@@ -12,7 +12,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.upgrad.FoodOrderingApp.service.business.CustomerService;
+import com.upgrad.FoodOrderingApp.service.businness.CustomerService;
 import com.upgrad.FoodOrderingApp.service.entity.CustomerAuthEntity;
 import com.upgrad.FoodOrderingApp.service.entity.CustomerEntity;
 import com.upgrad.FoodOrderingApp.service.exception.AuthenticationFailedException;
@@ -29,7 +29,6 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
-
 // This class contains all the test cases regarding the customer controller
 @RunWith(SpringRunner.class)
 @SpringBootTest
@@ -37,12 +36,9 @@ import org.springframework.test.web.servlet.MockMvc;
 public class CustomerControllerTest {
     @Autowired
     private MockMvc mockMvc;
-
     @MockBean
     private CustomerService mockCustomerService;
-
     // ----------------------------- POST /customer/signup --------------------------------
-
     //This test case passes when you are able to signup successfully.
     @Test
     public void shouldSignUpForValidRequest() throws Exception {
@@ -50,7 +46,6 @@ public class CustomerControllerTest {
         final String customerId = UUID.randomUUID().toString();
         createdCustomerEntity.setUuid(customerId);
         when(mockCustomerService.saveCustomer(any())).thenReturn(createdCustomerEntity);
-
         mockMvc
                 .perform(post("/customer/signup")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -59,28 +54,22 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("id").value(customerId));
         verify(mockCustomerService, times(1)).saveCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to signup but the request field is empty.
     @Test
     public void shouldNotSignUpForEmptyRequest() throws Exception {
-        when(mockCustomerService.saveCustomer(any()))
-            .thenThrow(new SignUpRestrictedException("SGR-005", "Except last name all fields should be filled"));
-
         mockMvc
                 .perform(post("/customer/signup")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
                         .content("{\"first_name\":\"first\", \"last_name\":\"last\", \"email_address\":\"\", \"contact_number\":\"9090909090\", \"password\":\"qawsedrf@123\"}"))
                 .andExpect(status().isBadRequest())
                 .andExpect(jsonPath("code").value("SGR-005"));
-       verify(mockCustomerService, times(1)).saveCustomer(any());
+        verify(mockCustomerService, times(0)).saveCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to signup with invalid email-id.
     @Test
     public void shouldNotSignUpForInvalidEmailId() throws Exception {
         when(mockCustomerService.saveCustomer(any()))
                 .thenThrow(new SignUpRestrictedException("SGR-002", "Invalid email-id format!"));
-
         mockMvc
                 .perform(post("/customer/signup")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -89,13 +78,11 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("SGR-002"));
         verify(mockCustomerService, times(1)).saveCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to signup with invalid contact number.
     @Test
     public void shouldNotSignUpForInvalidContactNo() throws Exception {
         when(mockCustomerService.saveCustomer(any()))
                 .thenThrow(new SignUpRestrictedException("SGR-003", "Invalid contact number!"));
-
         mockMvc
                 .perform(post("/customer/signup")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -104,13 +91,11 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("SGR-003"));
         verify(mockCustomerService, times(1)).saveCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to signup with invalid password.
     @Test
     public void shouldNotSignUpForInvalidPassword() throws Exception {
         when(mockCustomerService.saveCustomer(any()))
                 .thenThrow(new SignUpRestrictedException("SGR-004", "Weak password!"));
-
         mockMvc
                 .perform(post("/customer/signup")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -119,14 +104,12 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("SGR-004"));
         verify(mockCustomerService, times(1)).saveCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to signup with a contact number which is
     // already registered.
     @Test
     public void shouldNotSignUpIfTheContactIsAlreadySignedUp() throws Exception {
         when(mockCustomerService.saveCustomer(any()))
                 .thenThrow(new SignUpRestrictedException("SGR-001", "Try any other contact number, this contact number has already been taken"));
-
         mockMvc
                 .perform(post("/customer/signup")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -135,9 +118,7 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("SGR-001"));
         verify(mockCustomerService, times(1)).saveCustomer(any());
     }
-
     // ----------------------------- POST /customer/login --------------------------------
-
     //This test case passes when you are able to login successfully.
     @Test
     public void shouldLoginForValidRequest() throws Exception {
@@ -147,10 +128,8 @@ public class CustomerControllerTest {
         final String customerId = UUID.randomUUID().toString();
         customerEntity.setUuid(customerId);
         createdCustomerAuthEntity.setCustomer(customerEntity);
-
         when(mockCustomerService.authenticate("9090909090", "CorrectPassword"))
                 .thenReturn(createdCustomerAuthEntity);
-
         mockMvc
                 .perform(post("/customer/login")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -160,7 +139,6 @@ public class CustomerControllerTest {
                 .andExpect(header().exists("access-token"));
         verify(mockCustomerService, times(1)).authenticate("9090909090", "CorrectPassword");
     }
-
     //This test case passes when you have handled the exception of trying to login with invalid authorization format.
     @Test
     public void shouldNotLoginForInvalidAuthorizationFormat() throws Exception {
@@ -172,7 +150,6 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("ATH-003"));
         verify(mockCustomerService, times(0)).authenticate(anyString(), anyString());
     }
-
     //This test case passes when you have handled the exception of trying to login with a contact number that is not
     // registered yet.
     @Test
@@ -187,7 +164,6 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("ATH-001"));
         verify(mockCustomerService, times(1)).authenticate("123", "CorrectPassword");
     }
-
     //This test case passes when you have handled the exception of trying to login with incorrect password.
     @Test
     public void shouldNotLoginForWrongPassword() throws Exception {
@@ -201,10 +177,7 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("ATH-002"));
         verify(mockCustomerService, times(1)).authenticate("9090909090", "IncorrectPassword");
     }
-
-
     // ----------------------------- POST /customer/logout --------------------------------
-
     //This test case passes when you are able to logout successfully.
     @Test
     public void shouldLogoutForValidRequest() throws Exception {
@@ -214,7 +187,6 @@ public class CustomerControllerTest {
         customerEntity.setUuid(customerId);
         createdCustomerAuthEntity.setCustomer(customerEntity);
         when(mockCustomerService.logout("access-token")).thenReturn(createdCustomerAuthEntity);
-
         mockMvc
                 .perform(post("/customer/logout")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -223,13 +195,11 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("id").value(customerId));
         verify(mockCustomerService, times(1)).logout("access-token");
     }
-
     //This test case passes when you have handled the exception of trying to logout without even logging in.
     @Test
     public void shouldNotLogoutWhenCustomerIsNotLoggedIn() throws Exception {
         when(mockCustomerService.logout("auth"))
                 .thenThrow(new AuthorizationFailedException("ATHR-001", "Customer is not Logged in."));
-
         mockMvc
                 .perform(post("/customer/logout")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -238,13 +208,11 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("ATHR-001"));
         verify(mockCustomerService, times(1)).logout("auth");
     }
-
     //This test case passes when you have handled the exception of trying to logout when you have already logged out.
     @Test
     public void shouldNotLogoutIfCustomerIsAlreadyLoggedOut() throws Exception {
         when(mockCustomerService.logout("auth"))
                 .thenThrow(new AuthorizationFailedException("ATHR-002", "Customer is logged out. Log in again to access this endpoint."));
-
         mockMvc
                 .perform(post("/customer/logout")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -253,13 +221,11 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("ATHR-002"));
         verify(mockCustomerService, times(1)).logout("auth");
     }
-
     //This test case passes when you have handled the exception of trying to logout while your session is already expired.
     @Test
     public void shouldNotLogoutIfSessionIsExpired() throws Exception {
         when(mockCustomerService.logout("auth"))
                 .thenThrow(new AuthorizationFailedException("ATHR-003", "Your session is expired. Log in again to access this endpoint."));
-
         mockMvc
                 .perform(post("/customer/logout")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -268,9 +234,7 @@ public class CustomerControllerTest {
                 .andExpect(jsonPath("code").value("ATHR-003"));
         verify(mockCustomerService, times(1)).logout("auth");
     }
-
     // ----------------------------- PUT /customer --------------------------------
-
     //This test case passes when you are able to update customer details successfully.
     @Test
     public void shouldUpdateCustomerDetails() throws Exception {
@@ -278,9 +242,7 @@ public class CustomerControllerTest {
         customerEntity.setFirstName("firstname");
         final String customerId = UUID.randomUUID().toString();
         customerEntity.setUuid(customerId);
-
         when(mockCustomerService.getCustomer("auth")).thenReturn(customerEntity);
-
         final CustomerEntity updatedCustomerEntity = new CustomerEntity();
         updatedCustomerEntity.setFirstName("first");
         updatedCustomerEntity.setLastName("last");
@@ -296,7 +258,6 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(1)).getCustomer("auth");
         verify(mockCustomerService, times(1)).updateCustomer(customerEntity);
     }
-
     //This test case passes when you have handled the exception of trying to update user details but the first name
     // field is empty.
     @Test
@@ -311,14 +272,12 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(0)).getCustomer(anyString());
         verify(mockCustomerService, times(0)).updateCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to update customer details when the customer
     // is not logged in.
     @Test
     public void shouldNotUpdateCustomerDetailsWhenCustomerIsNotLoggedIn() throws Exception {
         when(mockCustomerService.getCustomer("auth"))
                 .thenThrow(new AuthorizationFailedException("ATHR-001", "Customer is not Logged in."));
-
         mockMvc
                 .perform(put("/customer")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -329,14 +288,12 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(1)).getCustomer("auth");
         verify(mockCustomerService, times(0)).updateCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to update customer details while you are
     // already logged out.
     @Test
     public void shouldUpdateCustomerDetailsIfCustomerIsAlreadyLoggedOut() throws Exception {
         when(mockCustomerService.getCustomer("auth"))
                 .thenThrow(new AuthorizationFailedException("ATHR-002", "Customer is logged out. Log in again to access this endpoint."));
-
         mockMvc
                 .perform(put("/customer")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -347,14 +304,12 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(1)).getCustomer("auth");
         verify(mockCustomerService, times(0)).updateCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to update customer details while your session
     // is already expired.
     @Test
     public void shouldUpdateCustomerDetailsIfSessionIsExpired() throws Exception {
         when(mockCustomerService.getCustomer("auth"))
                 .thenThrow(new AuthorizationFailedException("ATHR-003", "Your session is expired. Log in again to access this endpoint."));
-
         mockMvc
                 .perform(put("/customer")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -365,16 +320,13 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(1)).getCustomer("auth");
         verify(mockCustomerService, times(0)).updateCustomer(any());
     }
-
     // ----------------------------- PUT /customer/password --------------------------------
-
     //This test case passes when you are able to update your password successfully.
     @Test
     public void shouldUpdateCustomerPassword() throws Exception {
         final CustomerEntity customerEntity = new CustomerEntity();
         final String customerId = UUID.randomUUID().toString();
         customerEntity.setUuid(customerId);
-
         when(mockCustomerService.getCustomer("auth")).thenReturn(customerEntity);
         when(mockCustomerService.updateCustomerPassword("oldPwd", "newPwd", customerEntity))
                 .thenReturn(customerEntity);
@@ -388,7 +340,6 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(1)).getCustomer("auth");
         verify(mockCustomerService, times(1)).updateCustomerPassword("oldPwd", "newPwd", customerEntity);
     }
-
     //This test case passes when you have handled the exception of trying to update your password but your old password
     // field is empty.
     @Test
@@ -403,7 +354,6 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(0)).getCustomer(anyString());
         verify(mockCustomerService, times(0)).updateCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to update your password when your new password
     // field is empty
     @Test
@@ -418,14 +368,12 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(0)).getCustomer(anyString());
         verify(mockCustomerService, times(0)).updateCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to update your password but you are not
     // logged in.
     @Test
     public void shouldNotUpdateCustomerPasswordWhenCustomerIsNotLoggedIn() throws Exception {
         when(mockCustomerService.getCustomer("auth"))
                 .thenThrow(new AuthorizationFailedException("ATHR-001", "Customer is not Logged in."));
-
         mockMvc
                 .perform(put("/customer/password")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -436,13 +384,11 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(1)).getCustomer("auth");
         verify(mockCustomerService, times(0)).updateCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to update your password but you are already
     // logged out.
     public void shouldUpdateCustomerPasswordIfCustomerIsAlreadyLoggedOut() throws Exception {
         when(mockCustomerService.getCustomer("auth"))
                 .thenThrow(new AuthorizationFailedException("ATHR-002", "Customer is logged out. Log in again to access this endpoint."));
-
         mockMvc
                 .perform(put("/customer/password")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -453,13 +399,11 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(1)).getCustomer("auth");
         verify(mockCustomerService, times(0)).updateCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to update your password but your session is
     // already expired.
     public void shouldUpdateCustomerPasswordIfSessionIsExpired() throws Exception {
         when(mockCustomerService.getCustomer("auth"))
                 .thenThrow(new AuthorizationFailedException("ATHR-003", "Your session is expired. Log in again to access this endpoint."));
-
         mockMvc
                 .perform(put("/customer/password")
                         .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -470,7 +414,6 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(1)).getCustomer("auth");
         verify(mockCustomerService, times(0)).updateCustomer(any());
     }
-
     //This test case passes when you have handled the exception of trying to update your password while your new
     // password is weak.
     @Test
@@ -489,5 +432,4 @@ public class CustomerControllerTest {
         verify(mockCustomerService, times(1)).getCustomer("auth");
         verify(mockCustomerService, times(1)).updateCustomerPassword("oldPwd", "newPwd", customerEntity);
     }
-
 }
